+++ package.json
@@ -3,7 +3,7 @@
   "version": "2.0.5",
   "description": "The Local Notifications plugin allows your app to show notifications when the app is not running. Just like remote push notifications, but a few orders of magnitude easier to set up.",
   "main" : "local-notifications",
-  "typings": "local-notifications.d.ts",
+  "typings": "index.d.ts",
   "nativescript": {
     "platforms": {
       "android": "2.3.0",
@@ -17,6 +17,7 @@
   "keywords": [
     "ecosystem:NativeScript",
     "NativeScript",
+    "Alarm",
     "Notification",
     "Local Notification"
   ],
